search/requirements.txt
Django==4.2.19
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
sqlparse==0.5.3
PyYAML==5.4.1
urllib3==1.26.20
